ENH unique_classes is now a function of the multiclass module
