+++ sklearn/metrics/metrics.py
@@ -21,7 +21,7 @@ import warnings
 import numpy as np
 from scipy.sparse import coo_matrix
 
-from ..preprocessing import LabelBinarizer
+from ..multiclass import unique_labels
 from ..preprocessing import _is_label_indicator_matrix
 from ..utils import check_arrays, deprecated
 
@@ -87,32 +87,6 @@ def auc(x, y, reorder=False):
 
     area = np.sum(h * (y[1:] + y[:-1])) / 2.0
     return area
-
-
-def unique_labels(*lists_of_labels):
-    """Extract an ordered array of unique labels
-
-    Parameters
-    ----------
-    lists_of_labels : list of labels,
-        Each list of labels have to be in a format supported by
-        ``sklearn.preprocessing.LabelBinarizer``.
-
-    Returns
-    -------
-    out : numpy array of shape [n_unique_labels]
-        An ordered array of unique labels.
-
-    """
-    def _unique_labels(labels):
-        lb = LabelBinarizer()
-        lb.fit(labels)
-        return lb.classes_
-
-    if not lists_of_labels:
-        raise ValueError('No list of labels has been passed.')
-
-    return np.unique(np.hstack(_unique_labels(y) for y in lists_of_labels))
 
 
 ###############################################################################
